Tech debt work
