+++ hack/test-e2e-image.yaml
@@ -1,0 +1,49 @@
+apiVersion: v1
+kind: "Template"
+metadata:
+  name: elasticsearch-operator-test-build-template
+  annotations:
+    description: "Template for creating builds of elasticsearch operator test image."
+    tags: "infrastructure"
+labels:
+  logging-infra: test
+  provider: openshift
+  component: test
+objects:
+- apiVersion: v1
+  kind: ImageStream
+  metadata:
+    labels:
+      build: origin-elasticsearch-operator-e2e-test
+    name: origin-elasticsearch-operator-e2e-test
+  spec: {}
+- apiVersion: v1
+  kind: BuildConfig
+  metadata:
+    labels:
+      app: origin-elasticsearch-operator-e2e-test
+    name: origin-elasticsearch-operator-e2e-test
+  spec:
+    output:
+      to:
+        kind: ImageStreamTag
+        name: origin-elasticsearch-operator-e2e-test:latest
+    resources: {}
+    source:
+      git:
+        uri: ${ES_OP_GITHUB_URL}
+        ref: ${ES_OP_GITHUB_BRANCH}
+      type: Git
+    strategy:
+      dockerStrategy:
+        dockerfilePath: test/files/Dockerfile.centos7
+      type: Docker
+parameters:
+-
+  description: 'URL for elasticsearch-operator fork'
+  name: ES_OP_GITHUB_URL
+  value: https://github.com/openshift/elasticsearch-operator
+-
+  description: 'branch for elasticsearch-operator fork'
+  name: ES_OP_GITHUB_BRANCH
+  value: master
