+++ test/files/Dockerfile.centos7
@@ -1,0 +1,16 @@
+FROM openshift/origin-release:golang-1.10
+
+ENV HOME=${GOPATH}/src/github.com/openshift/elasticsearch-operator
+
+ADD . ${HOME}
+
+WORKDIR ${HOME}
+
+LABEL io.k8s.display-name="OpenShift elasticsearch-operator" \
+      io.k8s.description="This is the component that manages an Elasticsearch cluster on a kubernetes based platform" \
+      io.openshift.tags="openshift,logging,elasticsearch" \
+      com.redhat.delivery.appregistry=true \
+      maintainer="AOS Logging <aos-logging@redhat.com>"
+
+
+ENTRYPOINT ["./test/files/test-e2e.sh"]
