+++ test/files/e2e-test-pod.yaml
@@ -1,0 +1,28 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  labels:
+    run: elasticsearch-operator-e2e-test
+  name: elasticsearch-operator-e2e-test
+  namespace: ${TEST_NAMESPACE}
+spec:
+  containers:
+  - env:
+    - name: TEST_NAMESPACE
+      value: ${TEST_NAMESPACE}
+    - name: IMAGE_ELASTICSEARCH_OPERATOR
+      value: ${IMAGE_ELASTICSEARCH_OPERATOR}
+    image: ${IMAGE_E2E_TEST}
+    imagePullPolicy: Always
+    name: elasticsearch-operator-e2e-test
+    volumeMounts:
+    - mountPath: /tmp/test
+      name: kubeconfig
+  restartPolicy: Never
+  terminationGracePeriodSeconds: 30
+  volumes:
+  - name: kubeconfig
+    secret:
+      secretName: test-secret
+  command:
+  - test/files/test-e2e.sh
