+++ test/utils/utils.go
@@ -2,6 +2,8 @@ package utils
 
 import (
 	"io/ioutil"
+	"strconv"
+	"strings"
 	"testing"
 	"time"
 
@@ -10,7 +12,9 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
 
+	"github.com/openshift/elasticsearch-operator/pkg/k8shandler"
 	"github.com/openshift/elasticsearch-operator/pkg/utils"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -73,4 +77,131 @@ func GenerateUUID() string {
 	}
 
 	return uuid
+}
+
+func WaitForIndexTemplateReplicas(t *testing.T, kubeclient kubernetes.Interface, namespace, clusterName string, replicas int32, retryInterval, timeout time.Duration) error {
+	// mock out Secret response from client
+	mockClient := fake.NewFakeClient(getMockedSecret(clusterName, namespace))
+
+	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
+		// get all index replica count
+		indexTemplates, err := k8shandler.GetIndexTemplates(clusterName, namespace, mockClient)
+		if err != nil {
+			t.Logf("Received error: %v", err)
+			return false, nil
+		}
+
+		// for each index -- check replica count
+		for templateName, template := range indexTemplates {
+			if numberOfReplicas := parseString("settings.index.number_of_replicas", template.(map[string]interface{})); numberOfReplicas != "" {
+				currentReplicas, err := strconv.ParseInt(numberOfReplicas, 10, 32)
+				if err != nil {
+					return false, err
+				}
+
+				if int32(currentReplicas) == replicas {
+					continue
+				}
+
+				t.Logf("Index template %s did not have correct replica count (%d/%d)", templateName, currentReplicas, replicas)
+				return false, nil
+			} else {
+				return false, nil
+			}
+		}
+
+		return true, nil
+	})
+	if err != nil {
+		return err
+	}
+	t.Logf("All index templates have correct replica count of %d\n", replicas)
+	return nil
+}
+
+func WaitForIndexReplicas(t *testing.T, kubeclient kubernetes.Interface, namespace, clusterName string, replicas int32, retryInterval, timeout time.Duration) error {
+	// mock out Secret response from client
+	mockClient := fake.NewFakeClient(getMockedSecret(clusterName, namespace))
+
+	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
+
+		// get all index replica count
+		indexHealth, err := k8shandler.GetIndexReplicaCounts(clusterName, namespace, mockClient)
+		if err != nil {
+			return false, nil
+		}
+
+		// for each index -- check replica count
+		for index, health := range indexHealth {
+			if numberOfReplicas := parseString("settings.index.number_of_replicas", health.(map[string]interface{})); numberOfReplicas != "" {
+				currentReplicas, err := strconv.ParseInt(numberOfReplicas, 10, 32)
+				if err != nil {
+					return false, err
+				}
+
+				if int32(currentReplicas) == replicas {
+					continue
+				}
+
+				t.Logf("Index %s did not have correct replica count (%d/%d)", index, currentReplicas, replicas)
+				return false, nil
+			} else {
+				return false, nil
+			}
+		}
+
+		return true, nil
+	})
+	if err != nil {
+		return err
+	}
+	t.Logf("All indices have correct replica count of %d\n", replicas)
+	return nil
+}
+
+func getMockedSecret(clusterName, namespace string) *v1.Secret {
+	return Secret(
+		clusterName,
+		namespace,
+		map[string][]byte{
+			"elasticsearch.key": GetFileContents("test/files/elasticsearch.key"),
+			"elasticsearch.crt": GetFileContents("test/files/elasticsearch.crt"),
+			"logging-es.key":    GetFileContents("test/files/logging-es.key"),
+			"logging-es.crt":    GetFileContents("test/files/logging-es.crt"),
+			"admin-key":         GetFileContents("test/files/system.admin.key"),
+			"admin-cert":        GetFileContents("test/files/system.admin.crt"),
+			"admin-ca":          GetFileContents("test/files/ca.crt"),
+		},
+	)
+}
+
+func parseString(path string, interfaceMap map[string]interface{}) string {
+	value := walkInterfaceMap(path, interfaceMap)
+
+	if parsedString, ok := value.(string); ok {
+		return parsedString
+	} else {
+		return ""
+	}
+}
+
+func walkInterfaceMap(path string, interfaceMap map[string]interface{}) interface{} {
+
+	current := interfaceMap
+	keys := strings.Split(path, ".")
+	keyCount := len(keys)
+
+	for index, key := range keys {
+		if current[key] != nil {
+			if index+1 < keyCount {
+				current = current[key].(map[string]interface{})
+			} else {
+				return current[key]
+			}
+		} else {
+			return nil
+		}
+	}
+
+	return nil
 }
